move renderer next to the model
